Add an Artisan command that takes an argument for setting up the database in a specific state
